backend/requirements.txt

django>=4.2
djangorestframework
pandas
openpyxl
python-dotenv
# optional for OpenAI summary
openai